requirements.txt
python-telegram-bot==20.0
pillow==9.4.0
fpdf==1.7.2
python-dotenv==0.21.0
